Add reconnection support for Snowflake vector search.
