+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add redis plugin
 - Add pydantic schema support
 - Add recursive error propagation
+- Add reconnection support for Snowflake vector search.
 
 #### Bug Fixes
