+++ plugins/snowflake/superduper_snowflake/data_backend.py
@@ -1,6 +1,4 @@
 import json
-import os
-import threading
 import time
 import typing as t
 import uuid
@@ -15,10 +13,8 @@ from superduper.backends.base.data_backend import BaseDataBackend
 from superduper.base.event import CreateTable
 from superduper.base.query import Query
 from superduper.base.schema import Schema
-from watchdog.events import FileSystemEventHandler
-from watchdog.observers import Observer
 
-from superduper_snowflake.connect import connect
+from superduper_snowflake.connect import connect, watch_token_file
 from superduper_snowflake.query import map_superduper_query_to_snowpark_query
 from superduper_snowflake.schema import superduper_to_snowflake_schema
 
@@ -37,37 +33,6 @@ INSERT INTO "{table}" ({columns})
 """
 
 
-db_lock = threading.Lock()
-SESSION_DIR = os.environ.get('SNOWFLAKE_SESSION_DIR') or '/snowflake/session'
-
-
-class _SnowflakeTokenWatcher(FileSystemEventHandler):
-    timeout = 60
-
-    def __init__(self, databackend):
-        super().__init__()
-        self.databackend = databackend
-
-    def on_any_event(self, event):
-        logging.warn(str(event))
-
-        if event.src_path.endswith('data_tmp') and event.event_type == 'moved':
-            with db_lock:
-                self.databackend.reconnect()
-
-
-def _watch_token_file(databackend):
-    observer = Observer()
-    handler = _SnowflakeTokenWatcher(databackend)
-
-    logging.info(f'Starting Snowflake token watcher on {SESSION_DIR}/token')
-
-    observer.schedule(handler, SESSION_DIR, recursive=False)
-    observer.start()
-    logging.info('Started Snowflake token watcher')
-    return observer
-
-
 class SnowflakeDataBackend(BaseDataBackend):
     """Snowflake data backend."""
 
@@ -78,7 +43,7 @@ class SnowflakeDataBackend(BaseDataBackend):
         self.session, self.schema = connect(uri)
         self.observer = None
         if self.uri == 'snowflake://':
-            self.observer = _watch_token_file(self)
+            self.observer = watch_token_file(self)
 
     def reconnect(self):
         """Reconnect to the data backend."""
